change number integer to float

+++ db/migrate/20170114095601_add_number_to_food_recipe.rb
@@ -1,5 +1,5 @@
 class AddNumberToFoodRecipe < ActiveRecord::Migration[5.0]
   def change
-    add_column :food_recipes, :number, :integer, after: :id
+    add_column :food_recipes, :number, :float, after: :id
   end
 end
